Move utility functions out of LayerSprite and into point math
