+++ src/rendering/canvas-elements/layer-sprite.ts
@@ -35,7 +35,7 @@ import type { CanvasContextPairing, CanvasDrawable, Brush, BrushToolOptions, Bru
 import { LayerTypes } from "@/definitions/layer-types";
 import ToolTypes, { canDrawOnSelection } from "@/definitions/tool-types";
 import { scaleRectangle, rotateRectangle } from "@/math/rectangle-math";
-import { translatePointerRotation } from "@/math/point-math";
+import { translatePointerRotation, rotatePointer } from "@/math/point-math";
 import { renderEffectsForLayer } from "@/services/render-service";
 import { getBlendContext, blendLayer } from "@/rendering/blending";
 import { clipContextToSelection } from "@/rendering/clipping";
@@ -311,9 +311,8 @@ class LayerSprite extends ZoomableSprite {
         let selection: Selection = optAction?.selection || this._selection;
 
         // get the enqueued pointers which are to be rendered in this paint cycle
-        const pointers = isDrawing ? sliceBrushPointers( this._brush ) : undefined;
-        // @todo these are not pointers...
-        const transformedPointers = createOverrideConfig( this.canvas, pointers );
+        const pointers  = isDrawing ? sliceBrushPointers( this._brush ) : undefined;
+        const overrides = createOverrideConfig( this.canvas, pointers );
         
         // most drawing operations operate directly onto a temporary Canvas
         const usePaintCanvas = this.usePaintCanvas() || ( optAction?.type === "stroke" );
@@ -327,7 +326,7 @@ class LayerSprite extends ZoomableSprite {
             if ( selection ) {
                 ctx.save();
                 // note no offset is required when drawing on the full-size _paintCanvas
-                clipContextToSelection( ctx, selection, 0, 0, this._invertSelection, transformedPointers );
+                clipContextToSelection( ctx, selection, 0, 0, this._invertSelection, overrides );
             }
         } else if ( selection ) {
             ctx.save();
@@ -358,9 +357,9 @@ class LayerSprite extends ZoomableSprite {
             }
         } else if ( isDrawing ) {
             if ( isCloneStamp ) {
-                this._lastBrushIndex = renderClonedStroke( ctx, this._brush, this, this.toolOptions.sourceLayerId, pointers, transformedPointers, this._lastBrushIndex );
+                this._lastBrushIndex = renderClonedStroke( ctx, this._brush, this, this.toolOptions.sourceLayerId, pointers, overrides, this._lastBrushIndex );
             } else {
-                this._lastBrushIndex = renderBrushStroke( ctx, this._brush, transformedPointers, this._lastBrushIndex );
+                this._lastBrushIndex = renderBrushStroke( ctx, this._brush, overrides, this._lastBrushIndex );
             }
         }
         
@@ -743,35 +742,6 @@ class LayerSprite extends ZoomableSprite {
 export default LayerSprite;
 
 /* internal non-instance methods */
-
-function rotatePointers( pointers: Point[], layer: Layer, sourceWidth: number, sourceHeight: number ): Point[] {
-    // we take layer.left instead of bounds.left as it provides the unrotated Layer offset
-    const { left, top } = layer;
-    // translate pointer to translated space, when layer is rotated or mirrored
-    const { mirrorX, mirrorY, rotation } = layer.effects;
-    return pointers.map( point => {
-        // translate recorded pointer towards rotated point
-        // and against layer position
-        const p = translatePointerRotation(
-            point.x - left,
-            point.y - top,
-            sourceWidth  * HALF,
-            sourceHeight * HALF,
-            mirrorY ? -rotation : rotation
-        );
-        if ( mirrorX ) {
-            p.x -= sourceWidth;
-        }
-        if ( mirrorY ) {
-            p.y -= sourceHeight;
-        }
-        return p;
-    });
-}
-
-function rotatePointer( x: number, y: number, layer: Layer, sourceWidth: number, sourceHeight: number ): Point {
-    return rotatePointers([{ x, y }], layer, sourceWidth, sourceHeight )[ 0 ];
-}
 
 // NOTE we use getSpriteForLayer() instead of passing the Sprite by reference
 // as it is possible the Sprite originally rendering the Layer has been disposed
